# invillia-store-service/src/main/java/com/invillia/models/Address.java
package com.invillia.models;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.JoinColumn;
import javax.persistence.ManyToOne;
import javax.persistence.NamedQueries;import javax.persistence.NamedQuery;
import javax.persistence.Table;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

@Entity
@Table(name = "address")
@NamedQueries({@NamedQuery(name = "Address.findAllByStoreId", query = "select a from Address a where a.store.id =:storeId order by a.address"),
	           @NamedQuery(name = "Address.findById", query = "select a from Address a where a.store.id =:storeId and a.id =:id"),
	           @NamedQuery(name = "Address.getAddressCountById", query = "select count(a) from Address a where a.id =:id")})
public class Address implements java.io.Serializable {

	private static final long serialVersionUID = 1L;
	@Id	
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	@Column(name = "id")
	private Long id;
	@Column(name = "address", length = 255, nullable = false)
	private String address;
	@Column (name = "city", length = 255)
	private String city;
	@Column(name = "state", length = 255)
	private String state;
	@Column(name = "postal_code")
	private String postalCode;	
	@JsonIgnoreProperties("addresses")
	@ManyToOne(optional = false)
	@JoinColumn(name = "store_id")
	private Store store;
	
	
	public Address() {
		
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getAddress() {
		return address;
	}

	public void setAddress(String address) {
		this.address = address;
	}

	public String getCity() {
		return city;
	}

	public void setCity(String city) {
		this.city = city;
	}

	public String getState() {
		return state;
	}

	public void setState(String state) {
		this.state = state;
	}

	public String getPostalCode() {
		return postalCode;
	}

	public void setPostalCode(String postalCode) {
		this.postalCode = postalCode;
	}
	
	public Store getStore() {
		return store;
	}

	public void setStore(Store store) {
		this.store = store;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((id == null) ? 0 : id.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Address other = (Address) obj;
		if (id == null) {
			if (other.id != null)
				return false;
		} else if (!id.equals(other.id))
			return false;
		return true;
	}	
}