# invillia-store-service/swagger/swagger-store-service.yaml
swagger: '2.0'

info:
  description: Official documentation for Invillia Store API
  version: 1.0.0
  title: Invillia Store Service
  termsOfService: http://localhost:8001/invillia/terms
  contact:
    name: Administrator
    url: http://localhost:8001/invillia/support
    email: administrator@invillia.com
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: localhost:8001
basePath: /invillia

tags:
- name: stores
  description: Everything about stores
  externalDocs:
    description: Find out more
    url: http://invillia.com/stores
- name: addresses
  description: Everything about addresses
  externalDocs:
    description: Find out more
    url: http://invillia.com/orders
    
schemes:
- http    

paths:
  /store:
    get:
      tags: 
        - stores
      summary:  get all stores 
      description: get all stores
      operationId: getStores
      produces: 
        - application/json
      responses:
        200:
          description: OK
          schema:
            items:
              $ref: '#/definitions/Store'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
            
    post:
      tags: 
        - stores
      summary: Create a new store
      description: Create a new store
      operationId: addStore
      consumes: 
        - application/json
      produces: 
        - application/json
      responses:
        201:
          description: OK
          schema:
            $ref: '#/definitions/Store'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Address'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Address'
            
  /store/{id}:
    get:
      tags: 
        - stores
      summary: Get a store
      description: Get a store from an id
      operationId: getStore
      produces: 
        - application/json
      parameters: 
        - name: id
          in: path
          type: integer
          format: int64
          required: true
        - name: store_
          in: body
          schema:
            $ref: '#/definitions/Store'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Store'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
    
    put:
      tags: 
        - stores
      summary: Update a store
      description:  Update a store
      operationId: updateStore
      consumes: 
        - application/json
      produces: 
        - application/json
      parameters: 
        - name: id
          in: path
          type: integer
          format: int64
          required: true
        - name: store_
          in: body
          schema:
            $ref: '#/definitions/Store'
      responses:
        204:
          description: No content
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorMessage'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:  
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
    
    delete:
      tags: 
        - stores
      summary: Delete a store
      description:  delete a store as long as does not has any order
      operationId: deleteStore
      parameters: 
        - name: id
          in: path
          type: integer
          format: int64
          required: true
        - name: store_
          in: body
          schema:
            $ref: '#/definitions/Store'
      responses:
        204:
          description: No content
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorMessage'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:  
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'   
            
  /order/{storeId}/address:
    get:
      tags: 
        - addresses
      summary: "Get all address from a store"
      description: "Get all address from a store" 
      operationId: "getAddresses"
      parameters: 
        - name: "storeId"
          in: "path"
          type: integer
          format: int64
          required: true
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Address'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
            
    post:
      tags: 
        - addresses
      summary: "Create a new address"
      description: "create a new address to an store"
      operationId: "addAddress"
      consumes: 
        - application/json
      produces: 
        - application/json
      parameters: 
        - name: "storeId"
          in: "path"
          type: integer
          format: int64
          required: true
        - name: "address"
          in: "body"
          required: true
          schema:
            $ref: '#/definitions/Address'
      responses:
        201:
          description: OK
          schema:
            $ref: '#/definitions/Address'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorMessage'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'  
            
  /order/{storeId}/address/{id}:
    get:
      tags: 
        - addresses
      summary: Get an address 
      description:  Get an address
      operationId: getAddress
      produces: 
        - application/json
      parameters: 
        - name: storeId
          in: path
          type: integer
          format: int64
          required: true
        - name: id
          in: path
          type: integer
          format: int64
          required: true  
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Address'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
            
    put:
      tags: 
        - addresses
      summary: Update an address
      description: Update an address
      operationId: updateAddress
      consumes: 
        - application/json
      produces: 
        - application/json
      parameters: 
        - name: storeId
          in: path
          type: integer
          format: int64
          required: true
        - name: id
          in: path
          type: integer
          format: int64
          required: true  
        - name: "address"
          in: "body"
          required: true
          schema:
            $ref: '#/definitions/Address'
      responses:
        204:
          description: No content
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorMessage'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'  
            
    delete:
      tags: 
        - addresses
      summary: Remove an address from an store
      description:  Remove an address from an store  
      operationId: deleteAddress
      parameters: 
        - name: storeId
          in: path
          type: integer
          format: int64
          required: true
        - name: id
          in: path
          type: integer
          format: int64
          required: true  
      responses:
        204:
          description: No content
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ErrorMessage'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorMessage'
            
definitions:
  Store:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: Store description
      addresses:
        $ref: '#/definitions/Address'  
        
  Address:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      address:
        type: string
        example: First Street, 1
      city:
        type: string
        example: "Los Angeles"
      state:
        type: string
        example: "California"
      postalCode:
        type: string
        example: "899556"
      store:
        $ref: '#/definitions/Store'    
        
  ErrorMessage:
    type: object
    properties:
      description:
        type: string
        example: An error has ocurred
      documentation:
        type: string
        example: http://invilia.com/store/service/errors/list/#field_required      
              
    